Move function analysis out of parsing.
Previously, when parsing the bindgen bindings we determined
how to materialize functions. This was architecturally weird
to do at parsing time. This commit adds an analysis phase later
where we do this.

This commit too is a net regression in complexity, lines of code,
etc. but subsequent commits should clean it up.
