+++ engine/src/conversion/analysis/mod.rs
@@ -13,5 +13,6 @@
 // limitations under the License.
 
 pub(crate) mod ctypes;
+pub(crate) mod fun;
 pub(crate) mod gc;
 pub(crate) mod pod; // hey, that rhymes
