+++ engine/src/conversion/parse/mod.rs
@@ -12,12 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod bridge_name_tracker;
-mod overload_tracker;
-pub(crate) mod parse_bindgen;
+mod parse_bindgen;
 mod parse_foreign_mod;
-mod rust_name_tracker;
 pub(crate) mod type_converter;
-mod unqualify;
 
 pub(crate) use parse_bindgen::ParseBindgen;
